% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bcoef.varshrinkest.R
\name{Bcoef.varshrinkest}
\alias{Bcoef.varshrinkest}
\title{Coefficient matrices (B coefficient) for an object of class 'varshrinkest'}
\usage{
Bcoef.varshrinkest(x)
}
\description{
A variant of vars::Bcoef(). Code is modified to avoid call to data matrices
($y, $datamat) and to use effective numbers of parameters of shrinkage
estimates.
}
