% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ_sh.R
\name{BQ_sh}
\alias{BQ_sh}
\title{BQ function for class "varshrinkest"}
\usage{
BQ_sh(x)
}
\arguments{
\item{x}{An object of class "varshrinkest" obtained by VARshrink().}
}
\description{
An extension of vars::BQ() to the class "varshrinkest".
}
\examples{
data(Canada, package = "vars")
y <- diff(Canada)
estim <- VARshrink(y, p = 2, type = "const", method = "ridge")
BQ_sh(estim)
}
\seealso{
\code{\link[vars]{BQ}}
}
