% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.varshrinkest.R
\name{fevd.varshrinkest}
\alias{fevd.varshrinkest}
\title{Forecast Error Variance Decomposition}
\usage{
\method{fevd}{varshrinkest}(x, n.ahead = 10, ...)
}
\arguments{
\item{x}{Object of class "varshrinkest",
generated by \code{VARshrink()}.}

\item{n.ahead}{Integer specifying the steps.}

\item{...}{Currently not used.}
}
\value{
An object of class "varfevd" computed by \code{vars::fevd()}.
}
\description{
Computes the forecast error variance decomposition of a VAR(p) for
\code{n.ahead} steps.
}
\details{
It runs \code{vars::fevd()} for an object of class "varshrinkest".
}
\seealso{
\code{\link[vars]{fevd}}
}
