% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.varshrinkest.R
\name{predict.varshrinkest}
\alias{predict.varshrinkest}
\title{Predict method for objects of class varshrinkest}
\usage{
\method{predict}{varshrinkest}(object, n.ahead = 10, ci = 0.95, dumvar = NULL, ...)
}
\arguments{
\item{object}{An object of class "varshrinkest",
generated by \code{VARshrink()}}

\item{n.ahead}{An integer specifying the number of forecast steps.}

\item{ci}{The forecast confidence interval}

\item{dumvar}{Matrix for objects of class "vec2var" or "varest",
if the \code{dumvar} argument in \code{ca.jo()} has been used or if the
exogen argument in \code{VARshrink()} has been used, respectively. The matrix
should have the same column dimension as in the call to \code{ca.jo()} or to
\code{VARshrink()} and row dimension equal to n.ahead.}

\item{...}{Currently not used.}
}
\value{
A list with class attribute "varprd", containing the following
elements: \code{fcst}, \code{endog}, \code{model}, \code{exo.fcst}.
}
\description{
Forecating a VAR object of class "varshrinkest" with confidence bands.
}
\details{
This is a modification of \code{vars::predict.varest()} for the class
"varshrinkest".
}
\seealso{
\code{\link[vars]{predict.varest}}
}
