% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varshrinkest.R
\name{print.varshrinkest}
\alias{print.varshrinkest}
\title{Print method for class "varshrinkest"}
\usage{
\method{print}{varshrinkest}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "varshrinkest"}

\item{digits, ...}{Other arguments for \code{print()} method}
}
\description{
\code{print} method for an object of class "varshrinkest"
}
\examples{
data(Canada, package = "vars")
y <- diff(Canada)
estim <- VARshrink(y, p = 2, type = "const", method = "ridge")
print(estim)
}
