% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varshsum.R
\name{print.varshsum}
\alias{print.varshsum}
\title{Print method for class "varshsum"}
\usage{
\method{print}{varshsum}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class "varshsum"}

\item{digits, signif.stars, ...}{Other arguments for \code{print()},
\code{printCoefmat()}, \code{format()} method}
}
\description{
\code{print} method for an object obtained by \code{summary.varshrinkest()}.
}
\details{
This function extends \code{print.varsum()} for VAR models estimated by
shrinkage methods.
The output includes the covariance matrix \code{Sigma} and the
degrees-of-freedom \code{dof} for the multivariate t-distribution for
the residuals.
}
