% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi.varshrinkest.R
\name{Psi.varshrinkest}
\alias{Psi.varshrinkest}
\title{Coefficient matrices of the orthogonalized MA representation}
\usage{
\method{Psi}{varshrinkest}(x, nstep = 10, ...)
}
\description{
Returns the estimated orthogonalized coefficient matrices of the moving
average representation of a stable VAR(p).
}
\details{
This is an extension of \code{vars::Psi.varest()} for the class
"varshrinkest".
}
