% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.varshrinkest.R
\name{roots.varshrinkest}
\alias{roots.varshrinkest}
\title{Eigenvalues of the companion coefficient matrix of
a VAR(p)-process}
\usage{
roots.varshrinkest(x, modulus = TRUE)
}
\description{
This is a variant of vars::roots() for an object of class 'varshrinkest',
VAR parameters estimated by VARshrink()
}
\details{
Extend the roots() to class 'varshrinest' to incorporate
adapted methods: Acoef.varshrinkest(), Bcoef.varshrink

Code is modified to avoid call to data matrices ($y, $datamat)
and to use effective numbers of parameters of shrinkage estimates.
}
