% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkVARcoef.R
\name{shrinkVARcoef}
\alias{shrinkVARcoef}
\title{Shrinkage Estimation of VAR coefficients based on Bayesian estimation}
\usage{
shrinkVARcoef(Y, X, lambda, dof = Inf, prior_type = "NCJ",
  TolDRes = 1e-04, m0 = ncol(Y))
}
\description{
Shrinkage Estimation of VAR coefficients based on Bayesian estimation
}
