% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVARmodel.R
\name{simVARmodel}
\alias{simVARmodel}
\title{Generate simulated multivariate time series data using the given VAR model}
\usage{
simVARmodel(numT, model, burnin = 0)
}
\description{
Generate simulated multivariate time series data using the given VAR model
}
