% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.shrinklm.R
\name{summary.shrinklm}
\alias{summary.shrinklm}
\title{Summarizing shrinkage estimates of an AR model}
\usage{
\method{summary}{shrinklm}(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{An object of class "shrinklm"}

\item{correlation}{If TRUE, the correlation matrix of the
the estimated coefficients is returned and printed.}

\item{symbolic.cor}{If TRUE, print the correlations in a symbolic form
rather than as numbers}

\item{...}{Currently not used.}
}
\value{
Returns a list with class attribute "summary.lm", which
contains summary statistics of the fitted linear model given in object.
}
\description{
\code{summary} method for an object of class "shrinklm".
}
\details{
The class "shrinklm" inherits from the class "lm", and extends it to
incorporate shrinkage estimates with the effective number of parameters.
}
