% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoef.varshrinkest.R
\name{Acoef.varshrinkest}
\alias{Acoef.varshrinkest}
\title{A variant of vars::Acoef() for an object of class 'varshrinkest'}
\usage{
Acoef.varshrinkest(x)
}
\description{
Code is modified to avoid call to data matrices ($y, $datamat)
and to use effective numbers of parameters of shrinkage estimates.
}
