% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch.test_sh.R
\name{arch.test_sh}
\alias{arch.test_sh}
\title{ARCH-LM test}
\usage{
arch.test_sh(x, lags.single = 16, lags.multi = 5,
  multivariate.only = TRUE)
}
\description{
An extension of vars::arch.test() to the class "varshrinkest".
Performs univariate and multivariate ARCH-LM tests for a VAR.
}
