% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSE_Acoef.R
\name{calcSSE_Acoef}
\alias{calcSSE_Acoef}
\title{Sum of squared error (SSE) between two VAR coefficients}
\usage{
calcSSE_Acoef(Acoef1, Acoef2)
}
\arguments{
\item{Acoef1, Acoef2}{list of coefficient matrices}
}
\value{
SSE value
}
\description{
VAR coefficients in form of a list of A matrices.
}
