% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convPsi2varresult.R
\name{convPsi2varresult}
\alias{convPsi2varresult}
\title{Convert Psi to varresult}
\usage{
convPsi2varresult(Psi, Y, X, lambda, scale_lambda = 1,
  type = c("const", "mean", "none"), ybar = NULL, xbar = NULL,
  Q_values = NULL, callstr = "")
}
\arguments{
\item{Psi}{M-by-K matrix Psi}

\item{Y}{N-by-K matrix}

\item{X}{N-by-M matrix}

\item{lambda}{a shrinkage intensity parameter, which will be used
for computing the effective number of parameters}

\item{scale_lambda}{a factor to be multiplied to lambda}

\item{type}{Either 'const', 'mean', or 'none', indicating the
constant type in the VAR shrinkage problem.}

\item{ybar}{global mean vector if type=='mean'; local mean if type=='const'}

\item{xbar}{local mean vector for datX if type=='const'}

\item{callstr}{The call to VARshrink()}
}
\value{
a list of 'lm' objects with components: coefficients,
residuals, fitted.values, rank, df.residual, call
}
\description{
Consider the VAR process, Y = X %*% Psi + E, in the matrix form.
This function returns a list of lm objects for the multiple
regression problems corresponding to each column of Y:
y_j = X %*% psi_j + e_j.
}
