% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convPsi2varresult.R
\name{convPsi2varresult}
\alias{convPsi2varresult}
\title{Convert Psi to varresult}
\usage{
convPsi2varresult(Psi, Y, X, lambda0, type = c("const", "trend", "both",
  "none"), ybar = NULL, xbar = NULL, Q_values = NULL, callstr = "")
}
\arguments{
\item{Psi}{M-by-K matrix Psi}

\item{Y}{N-by-K matrix}

\item{X}{N-by-M matrix}

\item{lambda0}{a rescaled shrinkage intensity parameter
to be used for computing the effective number of parameters}

\item{type}{the type of the constant vector in the VAR shrinkage problem,
which is one of "const", "trend", "both", or "none".}

\item{ybar, xbar}{mean vectors of Y and X which had been subtracted.}

\item{callstr}{The call to VARshrink()}

\item{Q-values}{either NULL or a nonnegative weights of length N,
used for X'QX and X'QY in the hat matrix}
}
\value{
a list of c('shrinklm', 'lm') objects with components:
coefficients, residuals, fitted.values, rank, df.residual, lambda0,
call, terms, svd
}
\description{
Consider the VAR process, Y = X %*% Psi + E, in the matrix form.
This function returns a list of 'shrinklm' objects, which is
a slight modification to the 'lm', for the multiple
regression problems corresponding to each column of Y:
y_j = X %*% psi_j + e_j.
}
\details{
Depending on the 'type', the fitted values (y_j) and the estimated
coefficients (psi_j) are adjusted automatically.
}
