% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.varshrinkest.R
\name{fevd.varshrinkest}
\alias{fevd.varshrinkest}
\title{Forecast Error Variance Decomposition}
\usage{
\method{fevd}{varshrinkest}(x, n.ahead = 10, ...)
}
\description{
Computes the forecast error variance decomposition of a VAR(p) for
n.ahead steps.
This is a modification of vars::fevd() for the class "varshrinkest".
}
\seealso{
\code{\link[vars]{fevd}}
}
