% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.varshrinkest.R
\name{irf.varshrinkest}
\alias{irf.varshrinkest}
\title{Impulse response function}
\usage{
\method{irf}{varshrinkest}(x, impulse = NULL, response = NULL,
  n.ahead = 10, ortho = TRUE, cumulative = FALSE, boot = TRUE,
  ci = 0.95, runs = 100, seed = NULL, ...)
}
\description{
Computes the impulse response coefficients of a VAR(p)
(or transformed VECM to VAR(p)) for n.ahead steps.
This is a modification of vars::irf() for the class "varshrinkest".
}
\seealso{
\code{\link[vars]{irf}}
}
