% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_semi_Bayes_PCV.R
\name{lm_semi_Bayes_PCV}
\alias{lm_semi_Bayes_PCV}
\title{Semiparametric Bayesian shrinkage estimation method for VAR models}
\usage{
lm_semi_Bayes_PCV(Y, X, dof = Inf, lambda = NULL, lambda_var = NULL,
  prior_type = c("NCJ", "CJ"), num_folds = 5, m0 = ncol(Y))
}
\description{
Semiparametric Bayesian shrinkage estimation method for VAR models
}
