% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ShVAR_KCV.R
\name{lm_ShVAR_KCV}
\alias{lm_ShVAR_KCV}
\title{Shrinkage estimation method for VAR models using K-fold cross validation}
\usage{
lm_ShVAR_KCV(Y, X, dof = Inf, lambda = NULL, lambda_var = 0,
  prior_type = c("NCJ", "CJ"), num_folds = 5, m0 = ncol(Y))
}
\description{
Shrinkage estimation method for VAR models using K-fold cross validation
}
