% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.varshrinkest.R
\name{logLik.varshrinkest}
\alias{logLik.varshrinkest}
\title{Log-likelihood method for an object of class 'varshrinkest',
VAR parameters estimated by VARshrink()}
\usage{
\method{logLik}{varshrinkest}(object, ...)
}
\description{
Returns the log-likelihood of an estimated VAR model
for class 'varshrinest'. Extend logLik.varest() to incorporate
1) multivariate t-dribution for residuals,
2) scale matrix Sigma provided by shrinkage methods, and
3) effective number of parameters provided by shrinkage methods.
}
