% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.varshrinkest.R
\name{logLik.varshrinkest}
\alias{logLik.varshrinkest}
\title{Log-likelihood method for class "varshrinkest"}
\usage{
\method{logLik}{varshrinkest}(object, ...)
}
\description{
Returns the log-likelihood of a VAR model estimated by VARshrink().
Extend vars::logLik.varest() to incorporate
1) multivariate t-dribution for residuals,
2) scale matrix Sigma provided by shrinkage methods, and
3) effective number of parameters provided by shrinkage methods.
}
