% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.test_sh.R
\name{normality.test_sh}
\alias{normality.test_sh}
\title{Normality, multivariate skewness and kurtosis test for VAR shrinkage estimates}
\usage{
normality.test_sh(x, ...)
}
\description{
An extension of vars::normallity.test() to the class "varshrinkest".
}
