% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.test_sh.R
\name{normality.test_sh}
\alias{normality.test_sh}
\title{Normality, multivariate skewness and kurtosis test for VAR shrinkage estimates}
\usage{
normality.test_sh(x, multivariate.only = TRUE)
}
\arguments{
\item{x}{An object of class "varshrinkest" obtained by VARshrink().}

\item{multivariate.only}{If TRUE, only the multivariate statistics
is computed.}
}
\description{
An extension of vars::normality.test() to the class "varshrinkest".
}
