% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi.varshrinkest.R
\name{Phi.varshrinkest}
\alias{Phi.varshrinkest}
\title{Coefficient matrices of the MA represention}
\usage{
\method{Phi}{varshrinkest}(x, nstep = 10, ...)
}
\arguments{
\item{x}{An object of class 'varshrinkest',
generated by \code{VARshrink()}.}

\item{nstep}{An integer specifying the number of moving error
coefficient matrices to be calculated.}

\item{...}{Currently not used.}
}
\description{
Returns the estimated coefficient matrices of the moving average
representation of a stable VAR(p), of an SVAR as an array or a converted
VECM to VAR.
This is a modification of vars::Phi() for the class "varshrinkest".
}
\seealso{
\code{\link[vars]{Phi}}
}
