% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.varshrinkest.R
\name{predict.varshrinkest}
\alias{predict.varshrinkest}
\title{Predict method for objects of class varshrinkest}
\usage{
\method{predict}{varshrinkest}(object, ..., n.ahead = 10, ci = 0.95, dumvar = NULL)
}
\arguments{
\item{object}{An object of class 'varshrinkest';
generated by \code{VARshrink()}}

\item{...}{currently not used.}

\item{n.ahead}{An integer specifying the number of forecast steps.}

\item{ci}{The forecast confidence interval}

\item{dumvar}{Matrix for objects of class ‘vec2var’ or ‘varest’,
if the dumvar argument in ca.jo() has been used or if the exogen
argument in VARshrink() has been used, respectively. The matrix should
have the same column dimension as in the call to ca.jo() or to
VARshrink() and row dimension equal to n.ahead.}
}
\description{
Forecating a VAR object of class 'varshrinkest'
with confidence bands.
This is a modification of vars::predict.varest() for the class
"varshrinkest".
}
