% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varshrinksum.R
\name{print.varshrinksum}
\alias{print.varshrinksum}
\title{Print method for an object of class 'varshrinksum',}
\usage{
\method{print}{varshrinksum}(x, digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\description{
Extend print.varsum() to output alternative results
including the scale matrix, Sigma, and degree-of-freedom, dof,
for multivariate t-distribution for residuals.
}
