% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varshsum.R
\name{print.varshsum}
\alias{print.varshsum}
\title{Print method for class "varshsum"}
\usage{
\method{print}{varshsum}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class "varshsum"}

\item{digits, signif.stars, ...}{Other arguments for print(),
printCoefmat(), format() method}
}
\description{
Print method for an object obtained by summary.varshrinkest().
}
\details{
This function extends print.varsum() for VAR models estimated by
shrinkage methods.
The output includes scale matrix Sigma and degree of freedom dof
for multivariate t-distribution for residuals.
}
