% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi.varshrinkest.R
\name{Psi.varshrinkest}
\alias{Psi.varshrinkest}
\title{Coefficient matrices of the orthogonalized MA representation}
\usage{
\method{Psi}{varshrinkest}(x, nstep = 10, ...)
}
\arguments{
\item{x}{An object of class "varshrinkest",
generated by \code{VARshrink()}.}

\item{nstep}{An integer specifying the number of orthogonalized moving
error coefficient matrices to be calculated.}

\item{...}{Currently not used.}
}
\value{
An array with dimension \eqn{K \times K \times (nstep + 1)} holding
the estimated coefficients of the orthogonalized moving average
representation.
}
\description{
Returns the estimated orthogonalized coefficient matrices of the moving
average representation of a stable VAR(p).
}
\details{
This is an extension of \code{vars::Psi.varest()} for the class
"varshrinkest".
}
\seealso{
\code{\link[vars]{Psi}}
}
