% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_sh.R
\name{restrict_sh}
\alias{restrict_sh}
\title{Restricted VAR}
\usage{
restrict_sh(x, ...)
}
\arguments{
\item{x}{An object of class "varshrinkest"}

\item{...}{Other arguments to vars::restrict()}
}
\description{
This is a modification of vars::restrict() for the class "varshrinkest".
Warning: THIS CODE IS NOT COMPLETE:
this function may raise an error because it ignores shrinkage
estimation.
}
\examples{
data(Canada, package = "vars")
y <- diff(Canada)
estim <- VARshrink(y, p = 2, type = "const", method = "ridge")
restrict_sh(estim)
}
\seealso{
\code{\link[vars]{restrict}}
}
