% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_sh.R
\name{restrict_sh}
\alias{restrict_sh}
\title{Restrict function}
\usage{
restrict_sh(x, ...)
}
\arguments{
\item{x}{An object of class "varshrinkest"}

\item{...}{Other arguments to vars::restrict()}
}
\description{
Warning: this function ignores shrinkage estimation, so it need to be
revised.
}
\examples{
data(Canada, package = "vars")
y <- diff(Canada)
estim <- VARshrink(y, p = 2, type = "const", method = "ridge")
restrict_sh(estim)
}
\seealso{
\code{\link[vars]{restrict}}
}
