% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.test_sh.R
\name{serial.test_sh}
\alias{serial.test_sh}
\title{Test for serially correlated errors for VAR shrinkage estimate}
\usage{
serial.test_sh(x, ...)
}
\description{
An extension of vars::serial.test() to the class "varshrinkest".
}
