% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_sh.R
\name{stability_sh}
\alias{stability_sh}
\title{Stability function}
\usage{
stability_sh(x, type = c("OLS-CUSUM", "Rec-CUSUM", "Rec-MOSUM",
  "OLS-MOSUM", "RE", "ME", "Score-CUSUM", "Score-MOSUM", "fluctuation"),
  h = 0.15, dynamic = FALSE, rescale = TRUE, ...)
}
\description{
Warning: this function has not been tested for small sample sizes yet.
}
