% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_sh.R
\name{stability_sh}
\alias{stability_sh}
\title{Stability function}
\usage{
stability_sh(x, type = c("OLS-CUSUM", "Rec-CUSUM", "Rec-MOSUM",
  "OLS-MOSUM", "RE", "ME", "Score-CUSUM", "Score-MOSUM", "fluctuation"),
  h = 0.15, dynamic = FALSE, rescale = TRUE, ...)
}
\arguments{
\item{x}{An object of class "varshrinkest"}

\item{type, h, dynamic, rescale, ...}{Other arguments to strucchange::efp()}
}
\description{
A variant of vars::stability().
Warning: this function has not been tested for small sample sizes yet.
}
\examples{
data(Canada, package = "vars")
y <- diff(Canada)
estim <- VARshrink(y, p = 2, type = "const", method = "ridge")
stability_sh(estim)
}
\seealso{
\code{\link[vars]{stability}}
}
