% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.shrinklm.R
\name{summary.shrinklm}
\alias{summary.shrinklm}
\title{Summary method for class "shrinklm"}
\usage{
\method{summary}{shrinklm}(object, correlation = FALSE,
  symbolic.cor = FALSE, ...)
}
\description{
Class "shrinklm" inherits the class "lm", and it extends
the "lm" class to incorporate shrinkage estimates with
effective number of parameter.
}
