% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.shrinklm.R
\name{summary.shrinklm}
\alias{summary.shrinklm}
\title{Summary method for class "shrinklm"}
\usage{
\method{summary}{shrinklm}(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{An object of class "shrinklm"}

\item{correlation}{If TRUE, the correlation matrix of the
the estimated coefficients is returned and printed.}

\item{symbolic.cor}{If TRUE, print the correlations in a symbolic form
rather than as numbers}

\item{...}{Currently not used.}
}
\description{
Class "shrinklm" inherits the class "lm", and it extends
the "lm" class to incorporate shrinkage estimates with
effective number of parameter.
}
