% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARshrink.R
\name{VARshrink}
\alias{VARshrink}
\title{Shrinkage estimation of VAR parameters}
\usage{
VARshrink(y, p = 1, type = c("const", "mean", "none"),
  method = c("ridge", "ns", "fbayes", "sbayes", "kcv"), lambda = NULL,
  lambda_var = NULL, dof = Inf, ...)
}
\arguments{
\item{y}{T-by-K time series data}

\item{p}{lag order}
}
\description{
Shrinkage estimation of VAR parameters
}
